ocr-service/pyproject.toml
[project]
name = "ocr-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mlx>=0.31.0",
    "mlx-vlm>=0.3.12",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pypdfium2>=5.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "robyn>=0.79.0",
]
